Succesfull aggregate deprecation
